+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 🐞 Fixes
 
+- Fixed a glob input parsing issue when the pattern contained a `?` and wasn't in URI format.
 - Fixed JSON schemas not including property aliases (project `type` for `layer`, etc).
 
 #### ⚙️ Internal